@@ -1,32 +1,32 @@
 {
-    "name": "react-redux-realworld-example-app",
+    "name": "react-korek-frontend",
     "version": "0.1.0",
     "private": true,
     "devDependencies": {
-        "cross-env": "^5.2.0",
-        "react-scripts": "^3.0.0"
+        "cross-env": "^6.0.0",
+        "react-scripts": "^3.1.2"
     },
     "dependencies": {
         "bootstrap": "^4.3.1",
-        "history": "^4.9.0",
+        "history": "^4.10.1",
         "leaflet": "^1.5.1",
-        "rc-progress": "^2.3.0",
-        "react": "^16.8.6",
-        "react-bootstrap": "^1.0.0-beta.8",
-        "react-dom": "^16.3.0",
+        "rc-progress": "^2.5.2",
+        "react": "^16.9.0",
+        "react-bootstrap": "^1.0.0-beta.12",
+        "react-dom": "^16.9.0",
         "react-dropzone-component": "^3.2.0",
         "react-quill": "^1.3.3",
         "react-redux": "^5.0.7",
         "react-router": "^4.1.2",
         "react-router-dom": "^4.1.2",
         "react-router-redux": "^5.0.0-alpha.6",
-        "react-select": "^3.0.4",
-        "react-simple-tree-menu": "^1.1.4",
+        "react-select": "^3.0.5",
+        "react-simple-tree-menu": "^1.1.5",
         "react-widgets": "^4.4.11",
-        "redux": "^4.0.1",
-        "redux-devtools-extension": "^2.13.2",
-        "redux-logger": "^3.0.1",
-        "superagent": "^5.0.2",
+        "redux": "^4.0.4",
+        "redux-devtools-extension": "^2.13.8",
+        "redux-logger": "^3.0.6",
+        "superagent": "^5.1.0",
         "superagent-promise": "^1.1.0"
     },
     "scripts": {
